# ct-admin/src/main/resources/public/swagger/index.yaml
swagger: '2.0'
info:
  description: 是一套轻量级权限管理系统，极低门槛，拿来即用。设计之初，就非常注重安全性，为企业系统保驾护航，让一切都变得如此简单。
  version: 3.0.0
  title: ct-security
basePath: /ct-security
schemes:
  - http

#定义接口数据
paths:
...
  /sys/dept/info/{deptId}:
    get:
      tags:
        - 部门管理
      summary: 获取部门信息
      produces:
        - application/json
      parameters:
        - name: deptId
          description: 部门ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回部门信息
          schema:
            $ref: '#/definitions/SysDeptEntity'
  /sys/dept/save:
    post:
      tags:
        - 部门管理
      summary: 添加部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysDeptEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dept/update:
    post:
      tags:
        - 部门管理
      summary: 修改部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysDeptEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dept/delete:
    post:
      tags:
        - 部门管理
      summary: 删除部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/list:
    get:
      tags:
        - 角色管理
      summary: 角色列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: roleName
          description: 角色名
          in: query
          type: string
      responses:
        '200':
          description: 返回管理员列表
          schema:
            $ref: '#/definitions/SysRoleEntity'
...
  /sys/role/info/{roleId}:
    get:
      tags:
        - 角色管理
      summary: 获取角色信息
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            $ref: '#/definitions/SysRoleEntity'
  /sys/role/save:
    post:
      tags:
        - 角色管理
      summary: 添加角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/update:
    post:
      tags:
        - 角色管理
      summary: 修改角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/delete:
    post:
      tags:
        - 角色管理
      summary: 删除角色
      produces:
        - application/json
      parameters:
        - name: body
          description: 角色ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
        - 菜单管理
      summary: 获取菜单信息
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜单ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回菜单信息
          schema:
            $ref: '#/definitions/SysRoleEntity'
  /sys/menu/save:
    post:
      tags:
        - 菜单管理
      summary: 添加菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
        - 菜单管理
      summary: 修改菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete:
    post:
      tags:
        - 菜单管理
      summary: 删除菜单
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜单ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 状态码 0：成功  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  SysDeptEntity:
    type: object
    properties:
      deptId:
        description: 部门ID
        type: integer
        format: int64
      name:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门ID，一级部门为0
        type: integer
        format: int64
      parentName:
        description: 上级部门名称
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展开   true：展开   false：不展开
        type: boolean
        format: int32
  SysDeptEntityEdit:
    type: object
    properties:
      name:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门ID，一级部门为0
        type: integer
        format: int64
      orderNum:
        description: 排序
        type: integer
        format: int32
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysRoleEntityEdit:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
...
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
...